Put static functionality in ns_string! behind an unstable feature flag
