+++ crates/icrate/src/Foundation/__macro_helpers/cached.rs
@@ -1,0 +1,41 @@
+use core::mem::ManuallyDrop;
+use core::ptr;
+use core::sync::atomic::{AtomicPtr, Ordering};
+
+use objc2::rc::{Id, Shared};
+use objc2::Message;
+
+/// Allows storing an `Id` in a static and lazily loading it.
+pub struct CachedId<T> {
+    ptr: AtomicPtr<T>,
+}
+
+impl<T> CachedId<T> {
+    /// Constructs a new [`CachedId`].
+    pub const fn new() -> Self {
+        Self {
+            ptr: AtomicPtr::new(ptr::null_mut()),
+        }
+    }
+}
+
+impl<T: Message> CachedId<T> {
+    /// Returns the cached object. If no object is yet cached, creates one
+    /// from the given closure and stores it.
+    #[inline]
+    pub fn get(&self, f: impl FnOnce() -> Id<T, Shared>) -> &'static T {
+        // TODO: Investigate if we can use weaker orderings.
+        let ptr = self.ptr.load(Ordering::SeqCst);
+        // SAFETY: The pointer is either NULL, or has been created below.
+        unsafe { ptr.as_ref() }.unwrap_or_else(|| {
+            // "Forget" about releasing the object, effectively promoting it
+            // to a static.
+            let s = ManuallyDrop::new(f());
+            let ptr = Id::as_ptr(&s);
+            self.ptr.store(ptr as *mut T, Ordering::SeqCst);
+            // SAFETY: The pointer is valid, and will always be valid, since
+            // we haven't released it.
+            unsafe { ptr.as_ref().unwrap_unchecked() }
+        })
+    }
+}
