+++ crates/icrate/src/Foundation/__macro_helpers/mod.rs
@@ -1,3 +1,5 @@
+mod cached;
 mod ns_string;
 
+pub use self::cached::CachedId;
 pub use self::ns_string::*;
