+++ crates/test-ui/ui/invalid_ns_string_input.stderr
@@ -3,27 +3,3 @@ error[E0308]: mismatched types
   |
   |     let _ = ns_string!(1u8);
   |                        ^^^ expected `&str`, found `u8`
-
-note: erroneous constant used
- --> ui/invalid_ns_string_input.rs
-  |
-  |     let _ = ns_string!(1u8);
-  |             ^^^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/invalid_ns_string_input.rs
-  |
-  |     let _ = ns_string!(1u8);
-  |             ^^^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/invalid_ns_string_input.rs
-  |
-  |     let _ = ns_string!(1u8);
-  |             ^^^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
