+++ crates/test-ui/ui/ns_string_not_const.stderr
@@ -6,51 +6,3 @@ error[E0435]: attempt to use a non-constant value in a constant
   |             |          |
   |             |          non-constant value
   |             help: consider using `let` instead of `const`: `let INPUT`
-
-note: erroneous constant used
- --> ui/ns_string_not_const.rs
-  |
-  |     let _ = ns_string!(s);
-  |             ^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/ns_string_not_const.rs
-  |
-  |     let _ = ns_string!(s);
-  |             ^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/ns_string_not_const.rs
-  |
-  |     let _ = ns_string!(s);
-  |             ^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/ns_string_not_const.rs
-  |
-  |     let _ = ns_string!(s);
-  |             ^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/ns_string_not_const.rs
-  |
-  |     let _ = ns_string!(s);
-  |             ^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
-
-note: erroneous constant used
- --> ui/ns_string_not_const.rs
-  |
-  |     let _ = ns_string!(s);
-  |             ^^^^^^^^^^^^^
-  |
-  = note: this note originates in the macro `$crate::__ns_string_inner` which comes from the expansion of the macro `ns_string` (in Nightly builds, run with -Z macro-backtrace for more info)
